+++ src/CribblyBackend/Services/TeamService.cs
@@ -1,9 +1,6 @@
-using System.Collections.Generic;
-using System.Data;
-using System.Linq;
 using System.Threading.Tasks;
 using CribblyBackend.DataAccess.Models;
-using Dapper;
+using CribblyBackend.DataAccess.Repositories;
 
 namespace CribblyBackend.Services
 {
@@ -16,28 +13,16 @@ namespace CribblyBackend.Services
     }
     public class TeamService : ITeamService
     {
-        IDbConnection connection;
-        public TeamService(IDbConnection connection)
+        private readonly ITeamRepository _teamRepository;
+
+        public TeamService(ITeamRepository teamRepository)
         {
-            this.connection = connection;
+            _teamRepository = teamRepository;
         }
 
         public async Task Create(Team team)
         {
-            if (team.Players.Count < 2)
-            {
-                throw new System.Exception("A Team must not have less than two players");
-            }
-            await connection.ExecuteAsync(
-                @"INSERT INTO Teams(Name) VALUES (@Name)",
-                new { Name = team.Name }
-            );
-            foreach (Player player in team.Players)
-            {
-                await connection.ExecuteAsync(
-                    @"UPDATE Players SET TeamId = LAST_INSERT_ID() WHERE Id = @PlayerId",
-                    new { PlayerId = player.Id });
-            }
+            await _teamRepository.Create(team);
         }
 
         public void Delete(Team team)
@@ -47,22 +32,7 @@ namespace CribblyBackend.Services
 
         public async Task<Team> GetById(int id)
         {
-            var players = new Dictionary<int, Player>();
-            var team = (await connection.QueryAsync<Team, Player, Team>(
-                @"SELECT t.*, p.* FROM Teams t INNER JOIN Players p ON t.Id = p.TeamId WHERE p.TeamId = @Id",
-                (t, p) =>
-                {
-                    if (!players.TryGetValue(p.Id, out Player _))
-                    {
-                        players.Add(p.Id, p);
-                    }
-                    return t;
-                },
-                new { Id = id },
-                splitOn: "Id"
-                )).FirstOrDefault();
-            team.Players = players.Values.ToList();
-            return team;
+            return await _teamRepository.GetById(id);
         }
         public void Update(Team team)
         {
